addons/zalo/requirements.txt
python-zalo-bot
python-dotenv
httpx
fastapi
uvicorn